+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 * Add `Http.Headers.fromList`. In examples, this makes it easier
   to create HTTP requests.
 * Add `Eq` and `Show` instances for `Bodied`.
-* Get rid of `versionMajor` and `versionMinor`.
+* Get rid of `versionMajor` and `versionMinor` from both requests
+  and responses. Only support HTTP/1.1.
 
 ## 0.2.0.0 -- 2023-08-14